requirements.txt
tensorflow==2.15
numpy
pandas
matplotlib
seaborn
opencv-python-headless
scikit-learn
keras
flask
flask-cors
pillow